Update cmxDeployOnIbiza.py
If the requested CM credentials is admin then we do not need to create an admin user.

+++ cloudera-on-centos/scripts/cmxDeployOnIbiza.py
@@ -57,10 +57,12 @@ def init_cluster():
     :return:
     """
     #using default username/password to login first, create new admin user base on provided value, then delete admin
-    api = ApiResource(server_host=cmx.cm_server, username="admin", password="admin")
-    api.create_user(cmx.username, cmx.password, ['ROLE_ADMIN'])
-    api = ApiResource(server_host=cmx.cm_server, username=cmx.username, password=cmx.password)
-    api.delete_user("admin")
+    #if the requested cmcredential is admin then skip the below steps
+    if (cmx.username != "admin"):
+        api = ApiResource(server_host=cmx.cm_server, username="admin", password="admin")
+        api.create_user(cmx.username, cmx.password, ['ROLE_ADMIN'])
+        api = ApiResource(server_host=cmx.cm_server, username=cmx.username, password=cmx.password)
+        api.delete_user("admin")
 
     # Update Cloudera Manager configuration
     cm = api.get_cloudera_manager()
